Drop unread `Char::glyph_id` mapping from character clusters, improve docs (#688)
Similar to https://github.com/linebender/parley/pull/687, I believe
`parley` used to keep the glyph mapping around to pass to `swash`, but
`harfrust` does something like this internally.

<details>
<summary>This is a small but seemingly consistent speed up</summary>

```
$ cargo bench --bench main -- compare ../target/benchmarks/main -t 8.
Default Style - arabic 20 characters               [   8.8 us ...   8.8 us ]      -0.09%
Default Style - latin 20 characters                [   4.3 us ...   4.2 us ]      -1.80%*
Default Style - japanese 20 characters             [   8.2 us ...   8.1 us ]      -1.24%*
Default Style - arabic 1 paragraph                 [  47.7 us ...  47.1 us ]      -1.22%*
Default Style - latin 1 paragraph                  [  16.6 us ...  16.3 us ]      -1.63%*
Default Style - japanese 1 paragraph               [  70.4 us ...  69.0 us ]      -2.02%*
Default Style - arabic 4 paragraph                 [ 204.2 us ... 197.4 us ]      -3.32%*
Default Style - latin 4 paragraph                  [  62.7 us ...  61.4 us ]      -2.04%*
Default Style - japanese 4 paragraph               [  99.3 us ...  97.8 us ]      -1.57%*
Styled - arabic 20 characters                      [   9.8 us ...   9.8 us ]      -0.01%
Styled - latin 20 characters                       [   5.4 us ...   5.3 us ]      -1.98%*
Styled - japanese 20 characters                    [   8.7 us ...   8.6 us ]      -0.97%
Styled - arabic 1 paragraph                        [  50.2 us ...  49.7 us ]      -1.03%*
Styled - latin 1 paragraph                         [  21.2 us ...  20.8 us ]      -1.47%*
Styled - japanese 1 paragraph                      [  76.3 us ...  75.2 us ]      -1.45%*
Styled - arabic 4 paragraph                        [ 224.2 us ... 216.4 us ]      -3.50%*
Styled - latin 4 paragraph                         [  82.5 us ...  81.0 us ]      -1.87%*
Styled - japanese 4 paragraph                      [ 108.2 us ... 106.7 us ]      -1.40%*
```
</details>

---------

+++ parley/src/shape/mod.rs
@@ -257,12 +257,8 @@ impl<'a, 'b, B: Brush> FontSelector<'a, 'b, B> {
                     charmap
                         .map(ch)
                         .map(|g| {
-                            // HACK: in reality, we're only computing coverage, so
-                            // we only care about whether the font  has a mapping
-                            // for a particular glyph. Any non-zero value indicates
-                            // the existence of a glyph so we can simplify this
-                            // without a fallible conversion from u32 to u16.
-                            (g != 0) as u16
+                            // Any non-zero value indicates the existence of a glyph.
+                            g != 0
                         })
                         .unwrap_or_default()
                 },
